requirements.txt
requests
beautifulsoup4
cmd2
colorama
dnspython
asyncio
rich